+++ src/main/java/com/verdantartifice/primalmagic/common/sources/SourceList.java
@@ -15,6 +15,10 @@ import com.google.gson.JsonObject;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.ListTag;
 import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.network.chat.Component;
+import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraftforge.common.util.Constants;
 import net.minecraftforge.common.util.INBTSerializable;
 
@@ -253,5 +257,18 @@ public class SourceList implements INBTSerializable<CompoundTag> {
             }
         }
         return json;
+    }
+    
+    public Component getText() {
+        List<Source> contents = this.getSourcesSorted();
+        MutableComponent output = new TextComponent("");
+        for (int index = 0; index < contents.size(); index++) {
+            Source source = contents.get(index);
+            if (index != 0) {
+                output = output.append(new TextComponent(", "));
+            }
+            output = output.append(new TranslatableComponent("primalmagic.spells.details.mana_cost.piece", this.getAmount(source), source.getNameText()));
+        }
+        return output;
     }
 }